fix: castor refactor fixes Signed-off-by: Francisco Javier Ribo Labrador <elribonazo@gmail.com>
